refactor: fit standard go project layout
